% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcdl.R
\name{gcdl}
\alias{gcdl}
\title{Calculate P-values based on Graph-Constrained Desparsified LASSO (GCDL) Method}
\usage{
gcdl(x, y, G, nfolds = 10, Centering = TRUE)
}
\arguments{
\item{x}{The design matrix, which is an \code{n} by \code{p} matrix, where \code{n} is the number of observations and \code{p} is the number of predictors.}

\item{y}{The response vector, with \code{n} elements corresponding to the observations in the design matrix.}

\item{G}{User-specified graph structure matrix. \code{G[i, j]} indicates the presence of an edge between nodes \code{i} and \code{j}.}

\item{nfolds}{The number of cross-validation folds. Default is 10.}

\item{Centering}{Logical. Indicator of whether the design matrix should be centered to column zero mean. Default is \code{TRUE}.}
}
\value{
A list containing:
\itemize{
  \item \code{P_value} - Individual p-values for each parameter.
  \item \code{bhat} - The GCDL estimator.
  \item \code{betahat} - Initial estimate.
  \item \code{sigmahat} - The estimation of standard deviation obtained through the RCV method.
  \item \code{Se_bhat} - Individual standard deviation for each parameter.
  \item \code{Inv_Gram} - The approximate inverse of the matrix \eqn{x^\top x/n}.
}
}
\description{
Calculates P-values based on the graph-constrained desparsified LASSO (GCDL) method. This method incorporates the graph structure into the desparsified LASSO estimator, providing more accurate variable selection in high-dimensional settings.
}
\examples{
set.seed(0)
data <- simu_data(200, 20, 9)
x <- data$x
y <- data$y
G <- data$G
res <- gcdl(x, y, G)
print(res)
}
\references{
Chen, L., Liu, H., Kocher, J. P. A., Li, H., & Chen, J. (2015). glmgraph: an R package for variable selection and predictive modeling of structured genomic data. Bioinformatics, 31(24), 3991-3993.
}
