% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Laplacian.R
\name{Laplacian}
\alias{Laplacian}
\title{Calculate the Laplacian Matrix}
\usage{
Laplacian(G)
}
\arguments{
\item{G}{A square matrix representing the graph structure, where each entry \code{G[i, j]} 
indicates the weight of the edge between node \code{i} and node \code{j}. If \code{G[i, j]} is 
zero, it indicates that there is no edge between the nodes. Diagonal entries are ignored.}
}
\value{
A square matrix representing the Laplacian matrix \code{L}. The Laplacian matrix is 
calculated as \code{L = D - A}, where \code{D} is the degree matrix and \code{A} is the adjacency matrix.
}
\description{
This function computes the Laplacian matrix for a given graph structure matrix.
}
\details{
The adjacency matrix \code{A} is derived from the input graph structure matrix \code{G} by 
setting \code{A[i, j] = 1} if \code{G[i, j] != 0} and \code{A[i, j] = 0} otherwise. The degree matrix \code{D} 
is a diagonal matrix where each diagonal element \code{D[i, i]} is the sum of the corresponding row 
in the adjacency matrix \code{A}.
}
\examples{
set.seed(0)
data <- simu_data(200, 20, 9)
G <- data$G
lapmat <- Laplacian(G)
print(lapmat)

}
\seealso{
\code{\link{simu_data}}
}
