% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_estimator.R
\name{sd_estimator}
\alias{sd_estimator}
\title{Estimate Standard Deviation}
\usage{
sd_estimator(x, y, L, nfolds = 10)
}
\arguments{
\item{x}{The design matrix, which is a \code{n} by \code{p} matrix, where \code{n} is the number of observations and \code{p} is the number of predictors.}

\item{y}{The response vector, with \code{n} elements corresponding to the observations in the design matrix.}

\item{L}{Graph Laplacian matrix, which is used to incorporate the graph structure into the estimation process.}

\item{nfolds}{The number of cross-validation folds. Default is 10.}
}
\value{
The estimated standard deviation.
}
\description{
Utilizes the RCV method to estimate standard deviation. This method is particularly useful in high-dimensional settings where traditional variance estimation methods may not be effective. For more details on the RCV method, see Fan et al. (2012).
}
\examples{
set.seed(0)
data <- simu_data(200, 20, 9)
x <- data$x
y <- data$y
G <- data$G
lapmat <- Laplacian(G)
sd <- sd_estimator(x, y, lapmat, 10)
print(sd)
}
\references{
Fan, J., Guo, S., & Hao, N. (2012). Variance estimation using refitted cross-validation in ultrahigh dimensional regression. Journal of the Royal Statistical Society Series B: Statistical Methodology, 74(1), 37-65.
}
