% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_data.R
\name{simu_data}
\alias{simu_data}
\title{Generate the Simulation Data}
\usage{
simu_data(n, p, s0, Cov = NULL)
}
\arguments{
\item{n}{Integer. The sample size.}

\item{p}{Integer. The dimension of the covariates.}

\item{s0}{Integer. The cardinality of the active set. Must be a multiple of 3.}

\item{Cov}{A \code{p} by \code{p} positive definite covariance matrix. Default is \code{NULL}. If not provided, a default covariance matrix with a specific structure is generated.}
}
\value{
A list containing:
\itemize{
  \item \code{x} - The design matrix, where each row is an observation vector.
  \item \code{y} - The response vector.
  \item \code{G} - The graph structure matrix.
}
}
\description{
Generates simulation data based on a specified covariance structure. This function creates a design matrix, a response vector, and a graph structure matrix.
}
\examples{
set.seed(0)
data <- simu_data(200, 20, 9)
x <- data$x
y <- data$y
G <- data$G
}
